re-organize primary modules
