+++ web3/exceptions.py
